+++ src/components/Header.tsx
@@ -4,7 +4,7 @@ import StyledHeader from '@styles/components/Header.styled';
 
 export const Header = (): JSX.Element =>
   <StyledHeader>
-    <h6>TIMEWARD</h6>
+    <h6><Link to="/dashboard">TIMEWARD</Link></h6>
     <nav>
       <Link to="/dashboard">Dashboard</Link>
       <Link to="/login">Login</Link>